requirements.txt
# C-AIRA - Context-Aware Incident Response Assistant
# Core Dependencies

# AWS Bedrock SDK
boto3>=1.34.0
botocore>=1.34.0

# Vector Store and ML
faiss-cpu>=1.7.4
numpy>=1.24.0

# Text Processing
tiktoken>=0.5.2

# Web Framework
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
colorlog>=6.8.0
